3 more turrets and new playing concept
I've turned the game into a platformer where you can see from 1 side of the platform to anoter and is clamped .

Added 3 different turrets with unique shooting patterns.

Added a Game manager fot bullets that make a list and see how many bullets are in hierarchy . After a time, the bullet is disabled and  it is used again to shoot .

Next:
-3 more unique spells from the enemies
-wiggle the settings of the player
-player dash
-player charge power

PS: i had to buy a new charger for the laptop, the old one had burned up. Glad that i solved the problem