Added scoreboard and new turret
Added a new scoreboard system. The player need to move around the map and chase the PickUp boxes, one is +1, the rare one is +20.

Added a new turrete chat chase tour position and shoot projectiles based on a delay that can be adjusted.

Next: level gameplay and continous events to make it harder.