Update: added collision detection on the bullet for the player
Added a collision detection on the bullets for the player. When the bullets touch the player, the player dies.

Note: I've added code on the bullet not on the player just because i have some of them with Trigger and some of them with collision due to their patterns

I've changed the camera movement.Now you can rotate the camera around the circular platfoorm.

Next: Finish the 3 movement patterns of the Enemy:
-move in a square along the map
-move in random location of the map
-move towards player

This 3 patterns will change after some time (i want time not score or something else)